+++ src/TomorrowDAOServer.Application.Contracts/User/IUserAppService.cs
@@ -13,5 +13,4 @@ public interface IUserAppService
     Task<UserIndex> GetUserByCaHashAsync(string caHash);
     Task<string> GetUserAddressByCaHashAsync(string chainId, string caHash);
     Task<List<UserIndex>> GetUser();
-    Task<UserSourceReportResultDto> UserSourceReportAsync(string chainId, string source);
 }
