+++ src/TomorrowDAOServer.Application.Contracts/User/IUserService.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+using TomorrowDAOServer.User.Dtos;
+
+namespace TomorrowDAOServer.User;
+
+public interface IUserService
+{
+    Task<UserSourceReportResultDto> UserSourceReportAsync(string chainId, string source);
+}
